% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dgm_sample.R
\docType{data}
\name{dgm_sample}
\alias{dgm_sample}
\title{A random sample of a binary outcome MRT}
\format{
a data frame with 3000 observations and 10 variables
\describe{ This random sample uses the baseline model:
           log E(Y_{t+1} | A_t = 0, I_t = 1) = alpha_0 + alpha_1 * time / total_T
           + alpha_2 * 1(time > total_T/2),
           the treatment effect model:
           log relative risk = beta_0 + beta_1 * time / total_T,
           the probability of treatment assignment p_t: 0.3, 0.5, 0.7 with repetition,
           and exogenous probability of availability: 0.8 at all time points.
   \item{userid}{the subject IDs: from 1 to 100}
   \item{time}{time varibale that specifies decision points}
   \item{time_var1}{specfic decision point divided by total decision points}
   \item{time_var2}{indicates whether a specific time point has passed half of the total time points}
   \item{Y}{the binary outcome at all decision points: 0 = no outcome, 1 = outcome occurred}
   \item{A}{the binary treatment assigned at all decision points: 0 = no treatment, 1 = treatment}
   \item{avail}{the availability at all decision points}
   \item{prob_Y}{the probability of the binary outcome at decision points}
   \item{prob_Y_A0}{the probability of the binary outcome at the baseline treatment}
   \item{prob_A}{the probability of the treatment assignment at all decision points}
}
}
\usage{
dgm_sample
}
\description{
A random sample of a binary outcome MRT
}
\keyword{datasets}
