% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficient_ee.R
\name{efficient_ee}
\alias{efficient_ee}
\title{Estimates the conditional effect for binary outcome MRT}
\usage{
efficient_ee(
  dta,
  id_varname,
  decision_time_varname,
  treatment_varname,
  outcome_varname,
  control_varname,
  moderator_varname,
  rand_prob_varname,
  avail_varname = NULL,
  EMEE_initial_value = TRUE,
  estimator_initial_value = NULL
)
}
\arguments{
\item{dta}{a data set in long format}

\item{id_varname}{the variable name that specifies the subject IDs}

\item{decision_time_varname}{the variable name that specifies the decision points}

\item{treatment_varname}{the variable name that specifies the assigned treatments for subjects}

\item{outcome_varname}{the variable name that specifies the outcomes for subjects}

\item{control_varname}{a vector of variable names used to reduce noise, this could be NULL}

\item{moderator_varname}{a vector of variable names of the effect modifiers, this could be NULL}

\item{rand_prob_varname}{the variable name that specifies the treatment randomizing probability in the data set}

\item{avail_varname}{the variable name that specifies the availability of the subjects,
default to be always available at any decision points using NULL}

\item{EMEE_initial_value}{a logical value that specifies whether to use the estimates from
the marginal excursion effect estimator(estimator_EMEE()) as initial values,
default to be TRUE, which sets the initial values as the estimates from estimator_EMEE(),
use FALSE to avoid inputs from estimator_EMEE() if certain initial values
need to be specified in the parameter estimator_initial_value.}

\item{estimator_initial_value}{a numeric vector of the initial value for the estimator,
its length should be the sum of the length of control and moderator variables plus 2
default to be all 0's using NULL
the initial values in this estimator are very sensitive,
thus setting the initial values as the estimates from
the EMEE estimator is suggested.}
}
\value{
the estimated beta with its intercept and alpha with its intercept,
        the standard error of the estimated beta with its intercept and alpha with its intercept,
        the adjusted standard error of the estimated beta with its intercept and alpha with its intercept for small sample,
        the estimated variance-covariance matrix for the estimated beta with its intercept and alpha with its intercept,
        the estimated variance-covariance matrix for the estimated beta with its intercept and alpha with its intercept for small sample,
        the value of the estimating function at the estimated beta and alpha
}
\description{
This estimator returns the estimates for the semiparametric locally efficient estimator
and provides the estimated variance and standard error for the estimators,
with small sample correction for an MRT with binary outcome using the "Hat" matrix
in the variance estimate and t-distribution or F-distribution critical value with
corrected degrees of freedom.
}
\details{
This estimator should theoretically output approximately the same
results as the efficient_ee_twostep estimator and the efficient_ee_modified_weight
estimator. Please note that this estimator is very sensitive to the initial values,
and the other two estimators perform more stably than this one. Thus, please first
run the EMEE estimator and input those estimates as the initial values.
Alternatively, please try out the other two estimators.
}
\examples{
efficient_ee(dta = dgm_sample,
                       id_varname = "userid",
                       decision_time_varname = "day",
                       treatment_varname = "A",
                       outcome_varname = "Y",
                       control_varname = c("time_var1", "time_var2"),
                       moderator_varname = "time_var1",
                       rand_prob_varname = "prob_A",
                       EMEE_initial_value = TRUE)
}
