% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_change_location.R
\name{find_change_location}
\alias{find_change_location}
\title{Find the locations at which the value changes from the previous one}
\usage{
find_change_location(v)
}
\arguments{
\item{v}{a vector that specifies different values}
}
\value{
Return a vector that specifies the indexes of locations
        at which the value changes from the previous one
}
\description{
Find the locations at which the value changes from the previous one
}
\examples{
v <- c("a", "a", "b", "c", "c")
          find_change_location(v)
}
