% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alpha_beta_from_multiroot_result.R
\name{get_alpha_beta_from_multiroot_result}
\alias{get_alpha_beta_from_multiroot_result}
\title{Extract roots of alpha and beta from output of
the multiroot function}
\usage{
get_alpha_beta_from_multiroot_result(root, p, q)
}
\arguments{
\item{root}{a list of function output from the multiroot function}

\item{p}{a numeric value that specifies number of variables in beta}

\item{q}{a numeric value that specifies number of variables in alpha}
}
\value{
return a vector of roots for alpha,
        and a vector of roots for beta
}
\description{
Extract roots of alpha and beta from output of
the multiroot function
}
\examples{
f <- function(x) c(F1 = x[1] + x[2] - 1,
                             F2 = x[1] - x[2] - 7)
          solution <- rootSolve::multiroot(f = f, start = c(1, 1))
          get_alpha_beta_from_multiroot_result(root = solution,
                                              p = 1,
                                              q = 1)
}
